+++ language-service/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* Add `CucumberInfoBuilder` and `CucumberInfo`
+
 ### Changed
 
 ### Deprecated